chore: misc cleanups
