+++ apps/default-app/src/pages/AssetInventory/index.tsx
@@ -2,9 +2,7 @@ import { lazy } from "react";
 
 import { MainContainer } from "../../components/MainContainer";
 
-const AssetInventory = lazy(
-  () => import("../../../../../templates/AssetInventory"),
-);
+const AssetInventory = lazy(() => import("@hv/templates/AssetInventory"));
 
 const AssetInventoryWithContainer = () => (
   <MainContainer>
