+++ examples/uikit-app-shell/package.json
@@ -21,6 +21,7 @@
   },
   "devDependencies": {
     "@hitachivantara/app-shell-vite-plugin": "latest",
+    "@hitachivantara/uikit-config": "latest",
     "@hitachivantara/uikit-uno-preset": "latest",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
