+++ packages/core/src/Input/Input.test.tsx
@@ -3,8 +3,8 @@ import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { vi } from "vitest";
 
-import { HvInput, type HvInputProps } from ".";
 import { HvAdornment } from "../FormElement";
+import { HvInput, type HvInputProps } from "./Input";
 
 const Suggestions = ({ ...others }: Partial<HvInputProps>) => {
   const [value, setValue] = useState("");
